# docs/api/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Go Clean Architecture API",
        "title": "Go Clean Architecture API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "go-clean-architecture.xxxx.com",
    "basePath": "/",
    "paths": {
        "/api/user/register": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User Registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "User Registration Info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns access token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.GeneralSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/controller.GeneralErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Email already registered",
                        "schema": {
                            "$ref": "#/definitions/controller.GeneralErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.GeneralErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.GeneralErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controller.GeneralSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "request.Register": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8,
                    "example": "P@ssw0rd!"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}